refactor(mf): shutdown - scss into css
