--- micro-frontends/src/mf-shutdown/src/ui/modules/controller/controller.module.scss
@@ -1,6 +1,0 @@
-.controller {
-	display: flex;
-	column-gap: 5px;
-	justify-content: center;
-	padding-top: 35px;
-}
