+++ micro-frontends/src/mf-shutdown/src/ui/modules/description/description.module.css
@@ -1,0 +1,4 @@
+.description {
+	display: flex;
+	column-gap: var(--ds-spacing-md);
+}
