--- micro-frontends/src/mf-shutdown/src/ui/modules/description/description.module.scss
@@ -1,4 +1,0 @@
-.description {
-	display: flex;
-	column-gap: 10px;
-}
