+++ micro-frontends/src/mf-shutdown/src/ui/modules/presets/presets.module.css
@@ -1,7 +1,7 @@
 .presets {
 	display: flex;
 	flex-direction: column;
-	row-gap: 5px;
+	row-gap: var(--ds-spacing-sm);
 	margin-left: 42px;
 	margin-top: -10px;
 }
