--- micro-frontends/src/mf-shutdown/src/ui/modules/wrapper/wrapper.module.scss
@@ -1,6 +1,0 @@
-.wrapper {
-	max-width: 290px;
-	background-color: var(--background-color);
-	padding: 20px;
-	box-sizing: content-box;
-}
