+++ micro-frontends/src/mf-shutdown/src/ui/modules/wrapper/wrapper.tsx
@@ -2,7 +2,7 @@ import { Controller } from "../controller"
 import { Description } from "../description"
 import { Modal } from "../modal"
 import { Presets } from "../presets"
-import styles from "./wrapper.module.scss"
+import styles from "./wrapper.module.css"
 
 export const Wrapper = () => {
 	return (
